chore: simplify v2 — dead code removal, consistency fixes
Dead code removed:
- wal.go: packState and (*event).appendInt64 (unused helpers)
- wal.go: hasMsg field — production state read only by tests;
  setMessage ignores empty strings so hasMsg is equivalent to
  msg != ""; tests, fuzz model, and loom shadow updated
- wal_sim_test.go: step/sched/ownerActor (superseded by the op
  builders)
- operation_test.go: runPanic (never called)
- adapter/zap: errBoom test type, unreachable duplicate Err branch
  in fieldOf
- integration/std fuzz writer: hijackErr/hijackConn (unused fields)

Dead branches / ineffectual assignments:
- operation.go End: code/outcome initialized then unconditionally
  overwritten — assign directly from scanWAL/resolveOutcomeV2
- buildSampleInput: StatusCode re-assignment was a no-op (code is
  scan.code by construction)

Organization:
- copy.go merged into test_sink.go (its only consumer)
- policy.go merged into runtime.go beside validatePolicy/copyPolicy
- doc.go added: the package had no package-level doc comment

Consistency:
- go mod tidy: submodule go.sum files gain the root module's goleak
  entries (previously empty/stale)
- err := instead of var err error = (S1021/ST1023)
- fmt.Errorf instead of errors.New(fmt.Sprintf) (S1028)
- strings.Count(buf.String()) instead of string(buf.Bytes()) (S1030)

No behavior changes; full matrix (vet, test, -race, gofmt/gofumpt)
passes on every module.
