+++ runtime.go
@@ -174,6 +174,37 @@ func copyPolicy(policy OperationPolicy) OperationPolicy {
 	return policy
 }
 
+// levelFromPolicy resolves the final severity for an outcome under a
+// domain policy (zero fields mean the defaults).
+func levelFromPolicy(policy OperationPolicy, outcome Outcome) Level {
+	if outcomeLevel, ok := policy.OutcomeLevels[outcome]; ok && IsValidLevel(outcomeLevel) {
+		return outcomeLevel
+	}
+
+	// Zero policy fields mean the defaults; the zero Level is Info, which
+	// is the success default anyway. Explicit per-outcome control beyond
+	// that goes through OutcomeLevels.
+	successLevel, failureLevel, panicLevel := LevelInfo, LevelError, LevelError
+	if policy.SuccessLevel != 0 && IsValidLevel(policy.SuccessLevel) {
+		successLevel = policy.SuccessLevel
+	}
+	if policy.FailureLevel != 0 && IsValidLevel(policy.FailureLevel) {
+		failureLevel = policy.FailureLevel
+	}
+	if policy.PanicLevel != 0 && IsValidLevel(policy.PanicLevel) {
+		panicLevel = policy.PanicLevel
+	}
+
+	switch outcome {
+	case OutcomeSuccess:
+		return successLevel
+	case OutcomePanic:
+		return panicLevel
+	default:
+		return failureLevel
+	}
+}
+
 // noop reports whether the runtime can never emit (nil sink).
 func (rt *Runtime) noop() bool { return rt == nil || rt.sink == nil }
